+++ src/Router.php
@@ -31,12 +31,7 @@ final class Router
     #[\NoDiscard]
     public function __invoke(ServerRequest $request): Attempt
     {
-        return ($this->component)($request, SideEffect::identity())->mapError(
-            static fn($e) => match (true) {
-                $e instanceof Exception\HandleError => $e->unwrap(),
-                default => $e,
-            },
-        );
+        return ($this->component)($request, SideEffect::identity());
     }
 
     /**